making each db a test using specs/
